NHSO-17699: Add to calendar crashes if milliseconds supplied for start/end time

+++ xamarin/NHSOnline.App.Controls/WebViews/Payloads/AddEventToCalendarRequest.cs
@@ -6,8 +6,8 @@ namespace NHSOnline.App.Controls.WebViews.Payloads
             string subject,
             string body,
             string location,
-            int? startTimeEpochInSeconds,
-            int? endTimeEpochInSeconds)
+            decimal? startTimeEpochInSeconds,
+            decimal? endTimeEpochInSeconds)
         {
             Subject = subject;
             Body = body;
@@ -19,7 +19,7 @@ namespace NHSOnline.App.Controls.WebViews.Payloads
         public string Subject { get; set; }
         public string Body { get; set; }
         public string Location { get; set; }
-        public int? StartTimeEpochInSeconds { get; set; }
-        public int? EndTimeEpochInSeconds { get; set; }
+        public decimal? StartTimeEpochInSeconds { get; set; }
+        public decimal? EndTimeEpochInSeconds { get; set; }
     }
 }
