+++ xamarin/NHSOnline.App.iOS/Controllers/CalendarEventController.cs
@@ -70,8 +70,8 @@ namespace NHSOnline.App.iOS.Controllers
         private static EKEvent CreateCalendarEvent(AddEventToCalendarRequest request)
         {
             EKEvent calendarEvent = EKEvent.FromStore(ApplicationEventStore.Current.EventStore);
-            calendarEvent.StartDate = NSDate.FromTimeIntervalSince1970(request.StartTimeEpochInSeconds.GetValueOrDefault());
-            calendarEvent.EndDate = NSDate.FromTimeIntervalSince1970(request.EndTimeEpochInSeconds.GetValueOrDefault());
+            calendarEvent.StartDate = NSDate.FromTimeIntervalSince1970((long)request.StartTimeEpochInSeconds.GetValueOrDefault());
+            calendarEvent.EndDate = NSDate.FromTimeIntervalSince1970((long)request.EndTimeEpochInSeconds.GetValueOrDefault());
             calendarEvent.Title = request.Subject;
             calendarEvent.Notes = request.Body;
             calendarEvent.Location = request.Location;
